+++ frontend/src/app.tsx
@@ -1,9 +1,27 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { RouterProvider } from "@tanstack/react-router";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { RouterProvider } from "react-router";
 import { router } from "./router";
 import { CustomThemeProvider } from "./theme";
+
+// When navigating from this extension to another Fava page, Fava removes the node from the DOM.
+// However, React is still mounted in this (now detached) node.
+// Therefore, let's render React once into a detached node, and when the extension is visible, attach this node to the DOM tree.
+let appContainer: HTMLDivElement | null = null;
+export function renderApp(container: Element) {
+  if (!appContainer) {
+    appContainer = document.createElement("div");
+    const root = createRoot(appContainer);
+    root.render(
+      <StrictMode>
+        <App />
+      </StrictMode>,
+    );
+  }
+
+  container.replaceChildren(appContainer);
+}
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -13,15 +31,12 @@ const queryClient = new QueryClient({
   },
 });
 
-export function renderApp(container: Element) {
-  const root = createRoot(container);
-  root.render(
-    <StrictMode>
-      <QueryClientProvider client={queryClient}>
-        <CustomThemeProvider>
-          <RouterProvider router={router} />
-        </CustomThemeProvider>
-      </QueryClientProvider>
-    </StrictMode>,
+function App() {
+  return (
+    <QueryClientProvider client={queryClient}>
+      <CustomThemeProvider>
+        <RouterProvider router={router} />
+      </CustomThemeProvider>
+    </QueryClientProvider>
   );
 }